[refactor]: CacheMapper 함수 CacheState 위치로 이동 #61
